# api/openapi.yaml
openapi: 3.0.3
info:
  title: Fleetplane API
  description: >
    Agentless control plane for infrastructure fleets (fleetplane.io/v1alpha1).
    Every mutating request accepts an Idempotency-Key header; completed keys
    replay their stored response byte-identically.
  version: v1alpha1
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /v1/resources:
    post:
      operationId: createResource
      summary: Create a resource directly
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateResourceRequest" }
      responses:
        "201": { description: Resource journaled, content: { application/json: { schema: { $ref: "#/components/schemas/Resource" } } } }
        default: { $ref: "#/components/responses/Error" }
    get:
      operationId: listResources
      summary: List resources
      parameters:
        - { name: kind, in: query, schema: { type: string } }
        - { name: class, in: query, schema: { type: string } }
        - { name: provider, in: query, schema: { type: string } }
        - { name: pool, in: query, schema: { type: string } }
        - name: phase
          in: query
          description: repeatable
          style: form
          explode: true
          schema:
            type: array
            items: { type: string, enum: [unknown, provisioning, ready, allocated, draining, deleting, failed, orphaned, parking, parked, starting] }
      responses:
        "200": { description: Resources, content: { application/json: { schema: { $ref: "#/components/schemas/ResourceList" } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/resources/{id}:
    get:
      operationId: getResource
      summary: Read one resource
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Resource, content: { application/json: { schema: { $ref: "#/components/schemas/Resource" } } } }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteResource
      summary: Delete a resource (refused while leased/protected/unowned)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "202": { description: Deleting }
        default: { $ref: "#/components/responses/Error" }
  /v1/resources/{id}:drain:
    post:
      operationId: drainResource
      summary: Drain a resource (04 §3)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "202": { description: Draining }
        default: { $ref: "#/components/responses/Error" }
  /v1/resources/{id}:park:
    post:
      operationId: parkResource
      summary: Stop a ready machine into the parked tier (docs/12; 202 journaled, 200 idempotent no-op, 409 unsupported/conflict)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "202": { description: Parking journaled, content: { application/json: { schema: { type: object } } } }
        "200": { description: Already parking/parked, content: { application/json: { schema: { type: object } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/resources/{id}:start:
    post:
      operationId: startResource
      summary: Start a parked machine back into service (202 journaled, 200 idempotent no-op)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "202": { description: Start journaled, content: { application/json: { schema: { type: object } } } }
        "200": { description: Already starting/running, content: { application/json: { schema: { type: object } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/resources/{id}:undrain:
    post:
      operationId: undrainResource
      summary: Cancel a drain and return the resource to ready (200 — nothing is left in flight; already ready is 200, not 409)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Ready, content: { application/json: { schema: { type: object } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/resources/{id}:protect:
    post:
      operationId: protectResource
      summary: Set metadata.protected — refuses deletion, parking, idle reclaim and exclusive scheduling (200; already protected is 200)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Protected, content: { application/json: { schema: { type: object } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/resources/{id}:unprotect:
    post:
      operationId: unprotectResource
      summary: Clear metadata.protected (200; already unprotected is 200)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: Unprotected, content: { application/json: { schema: { type: object } } } }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, description: "flp_<id>.<secret> (ADR-008)" }
  parameters:
    id: { name: id, in: path, required: true, schema: { type: string } }
...
  responses:
    Error:
      description: Error envelope (single shape for every non-2xx)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            retryable: { type: boolean }
            retryAfterSeconds: { type: integer }
            details: { type: object }
    Metadata:
      type: object
      required: [id]
      properties:
        id: { type: string }
        name: { type: string }
        labels: { type: object, additionalProperties: { type: string } }
        ownership: { type: string, enum: [managed, adopted, observed] }
        protected: { type: boolean }
        generation: { type: integer }
        observedGeneration: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
    Resource:
      type: object
      required: [apiVersion, kind, metadata, spec]
      properties:
        apiVersion: { type: string }
        kind: { type: string }
        metadata: { $ref: "#/components/schemas/Metadata" }
        spec:
          type: object
          properties:
            kind: { type: string }
            provider: { type: string }
            class: { type: string }
            machine: { type: object }
        status:
          type: object
          properties:
            phase: { type: string, enum: [unknown, provisioning, ready, allocated, draining, deleting, failed, orphaned, parking, parked, starting] }
            externalId: { type: string }
            externalRef: { type: object }
            capacity: { type: object }
            extensions: { type: object, description: "provider-native, passed through untouched (invariant 6)" }
    ResourceList:
      type: object
      properties:
        apiVersion: { type: string }
        kind: { type: string }
        items: { type: array, items: { $ref: "#/components/schemas/Resource" } }
    CreateResourceRequest:
      type: object
      required: [spec]
      properties:
        apiVersion: { type: string }
        kind: { type: string }
        metadata: { $ref: "#/components/schemas/Metadata" }
        spec:
          type: object
          description: "Either set kind+provider with a full machine spec, or name a class: kind and provider are then inherited (a different value is a 400) and machine is a shallow overlay on the class spec — top-level keys such as image or userData replace the class value wholesale."
          properties:
            kind: { type: string }
            provider: { type: string }
            class: { type: string }
            machine: { type: object }
...